use package.json type
